Move fixtures to module
